// frontend/src/api/api/models/GameAddUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GameAddUpdate
 */
export interface GameAddUpdate {
    /**
     * 
     * @type {string}
     * @memberof GameAddUpdate
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof GameAddUpdate
     */
    introduction?: string;
    /**
     * 
     * @type {string}
     * @memberof GameAddUpdate
     */
    content?: string;
    /**
     * 
     * @type {Date}
     * @memberof GameAddUpdate
     */
    gamePublishDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof GameAddUpdate
     */
    developer?: string;
    /**
     * 
     * @type {number}
     * @memberof GameAddUpdate
     */
    editorScore?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof GameAddUpdate
     */
    genres?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof GameAddUpdate
     */
    platforms?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof GameAddUpdate
     */
    distributions?: Array<string>;
}

export function GameAddUpdateFromJSON(json: any): GameAddUpdate {
    return GameAddUpdateFromJSONTyped(json, false);
}

export function GameAddUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameAddUpdate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': !exists(json, 'title') ? undefined : json['title'],
        'introduction': !exists(json, 'introduction') ? undefined : json['introduction'],
        'content': !exists(json, 'content') ? undefined : json['content'],
        'gamePublishDate': !exists(json, 'gamePublishDate') ? undefined : (new Date(json['gamePublishDate'])),
        'developer': !exists(json, 'developer') ? undefined : json['developer'],
        'editorScore': !exists(json, 'editorScore') ? undefined : json['editorScore'],
        'genres': !exists(json, 'genres') ? undefined : json['genres'],
        'platforms': !exists(json, 'platforms') ? undefined : json['platforms'],
        'distributions': !exists(json, 'distributions') ? undefined : json['distributions'],
    };
}

export function GameAddUpdateToJSON(value?: GameAddUpdate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'introduction': value.introduction,
        'content': value.content,
        'gamePublishDate': value.gamePublishDate === undefined ? undefined : (value.gamePublishDate.toISOString()),
        'developer': value.developer,
        'editorScore': value.editorScore,
        'genres': value.genres,
        'platforms': value.platforms,
        'distributions': value.distributions,
    };
}